+++ packages/operator/pkg/servicecatalog/routes/healthcheck.go
@@ -14,16 +14,21 @@
  * limitations under the License.
  */
 
-package servicecatalog
+package routes
 
 import (
-	"net/http"
-
 	"github.com/gin-gonic/gin"
-	_ "github.com/odahu/odahu-flow/packages/operator/docs" //nolint
-	"github.com/odahu/odahu-flow/packages/operator/pkg/apiserver/routes"
+	"net/http"
 )
 
-func SetUpSwagger(rg *gin.RouterGroup, apiStaticFS http.FileSystem, reader routes.SwaggerDefinitionReader) {
-	rg.GET("/swagger/*any", routes.SwaggerHandler(apiStaticFS, reader))
+const (
+	HealthCheckURL = "/health"
+)
+
+func healthCheck(c *gin.Context) {
+	c.AbortWithStatusJSON(http.StatusOK, gin.H{})
+}
+
+func SetUpHealthCheck(server *gin.Engine) {
+	server.GET(HealthCheckURL, healthCheck)
 }
